pyproject.toml
[project]
name = "summaryagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "crawlee>=1.0.3",
    "playwright>=1.55.0",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
